+++ worker/walrus/src/models.rs
@@ -1,0 +1,3 @@
+pub mod realtime;
+pub mod wal2json;
+pub mod walrus;
